+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - cmd/lint-changelog: handle french date format.
 
+### Changed
+
+- validator: Improve error messages.
+
 ## [0.5.0] - 2025-10-28
 
 ### Changed